feat: make the entire logo branding a env dropdown selector
